doom update

+++ games/doom-nano_328p/doom-nano/entities.cpp
@@ -5,7 +5,7 @@
 
 Entity create_entity(uint8_t type, uint8_t x,  uint8_t y, uint8_t initialState, uint8_t initialHealth) {
   UID uid = create_uid(type, x, y);
-  Coords pos = create_coords((double) x + .5, (double) y + .5);
+  Coords pos = create_coords((float) x + .5, (float) y + .5);
   Entity new_entity = { uid, pos, initialState, initialHealth, 0, 0 };
   return new_entity;
 }
