+++ games/doom-nano_328p/doom-nano/types.h
@@ -20,14 +20,14 @@ typedef uint16_t UID;
 typedef uint8_t  EType;
 
 struct Coords {
-  double x;
-  double y;
+  float x;
+  float y;
 };
 
 UID create_uid(EType type, uint8_t x, uint8_t y);
 EType uid_get_type(UID uid);
 
-Coords create_coords(double x, double y);
+Coords create_coords(float x, float y);
 uint8_t coords_distance(Coords* a, Coords* b);
 
 #endif
